Move commands into command folder, changed "Send" to Instant Command
